Made the fxns necessary to solve algebraic equat
Made it so there is a step by step process to solve the algebraic functions; created simplify, isExpression, and solve. Looked into the future and also thought of a possible way to solve polynomials by checking if there is an exponent and creating an individual method to solve polynomials. Also Fixed the GUI